+++ app/src/main/java/com/gp/cryptotrackerapp/data/remote/model/CoinDataHistoryRemoteModel.kt
@@ -3,5 +3,5 @@ package com.gp.cryptotrackerapp.data.remote.model
 import android.util.Log
 
 data class CoinDataHistoryRemoteModel(
-    var prices : ArrayList<ArrayList<Int>>? = null
+    var prices : ArrayList<ArrayList<Double>>? = null
 )
